[update] devide data fetching from post list component
