+++ components/index/PostList.js
@@ -1,24 +1,10 @@
 import styles from "../../styles/index/PostList.module.css";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useModal } from "react-modal-hook";
 import PostModal from "../modals/PostModal";
 
-
 const PostList = props => {
-  // const posts = props.posts;
-   //TODO delete
-  const message = "アルパカ食べたいアルパカ食べたいアルパカ食べたいアルパカ食べたいアルパカ食べたいアルパカ食べたいアルパカ食べたいアルパカ食べたいアルパカ食べたい"
-  const post_list = [
-    {sentTo_1: "医療従事者", sentTo_2: "の方", message: "いつも私たちのために最前線で働いてくださり、ありがとうございます！"},
-    {sentTo_1: "コロナに", sentTo_2: "感染した方", message: "はやく治りますように！"},
-    {sentTo_1: "飲食店に", sentTo_2: "勤務の方", message: "いつも美味しいご飯をありがとうございます。"},
-    {sentTo_1: "アルパカ", sentTo_2: "の方", message: message},
-  ]
-  const posts = [...Array(20)].map((_, i) => {
-    const j = Math.floor(Math.random() * post_list.length);
-    return post_list[j];
-  });
-  // TODO -------------------------
+  const posts = props.posts;
 
   const [targetPost, setTargetPost] = useState("")
   const [readMoreCount, setReadMoreCount] = useState(1);
@@ -41,7 +27,6 @@ const PostList = props => {
     event.preventDefault();
     const target_node = event.currentTarget;
     const idx = Array.from(target_node.parentNode.children).indexOf(target_node);
-    console.log(idx);
     setTargetPost(posts[idx]);
     showModal();
   }
